Error instead of warning on illegal escape sequences.
The previous behavior for illegal escape sequences (e.g. '\%') in character
and string literals was to emit an error, then treat "\x" as "x".
As reported by a user, this is dangerous, because the warning can go
unnoticed, and other compilers can treat "\x" as "\\x"
(backslash followed by 'x').  Better to error out.