Relaxing the semantics of pointer arithmetic.
Support <pointer> +/- <integer> where the pointer value is actually an integer (Vint) that has been converted to pointer type.

Such arithmetic, while not defined in ISO C, appears in the wild.  If present in static initializers, it used to cause a compile-time failure ("not a compile-time constant").  Hence this relaxation.