Print floating-point numbers with more digits in debug outputs
As suggested in GPR#84, use '%.15F' to force the printing of more significant digits.  (The '%F' format previously used prints only 6.)  This is enough to represent the FP number exactly most of the time (but not always).

Once OCaml 4.03 is out and CompCert switches to this version of OCaml, we'll be able to use hexadecimal floats for printing.