Improved handling of C90 calls to undeclared functions
In a call such as "f(expr, ..., expr)", if the identifier "f" is not declared, declare it as specified in the ISO C90 standard, namely like "extern int f()" would do it.  Previously, the declaration was done "on the side" and not properly recorded in the environments.

The diff is relatively large because the "enter_or_refine_ident" had to be moved up in the file.