Regression: handling of integer + pointer in CompCert C
During the experiments, the integer + pointer cases was removed from the semantics of the C addition operator.  The idea was to turn integer + pointer into pointer + integer during elaboration, but it was not implemented.

On second thoughts, we can restore the integer + pointer cases in the formal semantics of CompCert C at low cost.  This is what this commit does.