Added support for -pipe.
Currently only the preprocessor is supported. This should work for
both the diab and gnu toolchain since in both cases the stdout of
the preprocessor was just redirected to a file.

The implementation works by introducing a new type that abstracts
away which kind of output is used for the preprocessing output.
Bug 20521