Close preprocessor earlier.
For the pipe case it can be the case that the preprocessor dies
and produces some output before. In order to avoid parsing the
unfinished preprocessed code producing parser errors and masking
the real error source the close_prepro_in is now performed
directly after reading in the whole file / pipe.

Furthermore for reading from files the old behavior of reading in
the whole file is restored.
Bug 20521