Implement offsetof via builtin.
The implementation of offsetof as macro in the form
((size_t) &((ty*) NULL)->member) has the problem that it cannot be
used everywhere were an integer constant expression is allowed,
for example in initiliazers of global variables and there is also
no check for the case that member is of bitifield type.

The new implementation adds a builtin function for this which is
replaced by an integer constant during elaboration.
Bug 20765