Added a simple check for unused variables.
The check test whether the identifier is used at all in the
function and if not issue a warning. It is not tested whether the
usage is reachable at all, so

int i;
if (0)
  i;

would not generate a warning. This is the same as gcc/clang does.

The warning is disabled per default, but is active if -Wall is
given.

Bug 19872