Tunnelingproof: Remove assumption destroyed_by_cond c = nil.
Since commit e5b37a6 (useless conditional branch elimination), the proof of the Tunneling pass was assuming forall c, destroyed_by_cond c = nil.  This is not true for architecture variants that we will support soon.  This commit rewrites the proof so as to remove this assumption.  The old proof was by memory and value equalities, the new one is by memory extensions and "lessdef" values.