ARM port: replace Psavelr pseudo-instruction by actual instructions
Saving the return address register (R14) in the function prologue can be done either with a single "str" instruction if the offset is small enough, or by constructing the address using the R12 register as a temporary then storing with "str" relative to R12.

R12 can be used as a temporary because it is marked as destroyed at function entry.  We just need to tell Asmgen.translcode whether R12 still contains the back link left there by Pallocframe, or R12 was trashed.  In the latter case R12 will be reloaded from the stack at the next Mgetparam instruction.