Deadcode: eliminate trivial Icond instructions
These are conditionals where the "ifso" and the "ifnot" successors are the same.  By eliminating them here and not later, we can also eliminate the instructions that compute the arguments to the condition, if any.

There is another, later point where these trivial conditional instructions are eliminated: in the Tunneling phase.  The elimination done in Tunneling is more powerful in that it works not just for conditionals whose two successors are the same, but also for conditionals whose two successors lead to the same point after a series of nops.  The elimination done in Deadcode is more powerful in that it eliminates the instructions that compute the arguments to the condition.  Hence it is worth having both eliminations.