x86 ConstpropOp.addr_strength_reduction: always check validity of resulting addressing
In the original code, the addressing_valid check is skipped if we are in 32 bits, because we know the check is always true.  This is correct but not obvious nor future-proof.  (In the future we may want to make addressing_valid more strict.)

This commit restructures ConstpropOp.addr_strength_reduction so that the addressing_valid check is always performed.