Extend the modorder tool to handle Coq files as well (#54)
This is useful to e.g. identify the .vo files from CompCert that a clightgen-generated .v file needs.

Also: the "result" field of the record type is now initialized with the LHS of the dependency, not the RHS.  It doesn't matter because the result field is unused, but it makes more sense now.