Drop uninterned symbols
Instead, we use regular symbols, but make sure that they are not legal
C identifiers, so that they cannot conflict with those parsed from the
source program. Specifically, the temporaries introduced by the
SimplExpr pass are of the form "$1", "$2", etc. (which is similar in
effect to the old behavior when they are printed out), and temporaries
introduced to separate memory loads in clightgen normalization are of
the form "@1", "@2", etc.