Make ident an abstract type
AST.ident values can now only be created through Ident.of_string, and
can only be used through Ident.to_string. This guarantees that the
concrete positive <-> string mapping used is immaterial and that ident
values can never be communicated with the outside world.

This discipline is enforced both in the Coq and Ocaml code: the
Camlcoq functions intern_string and extern_atom now use the sealed
Ident module (and type) as extracted from Coq. Independently, the
FastIdent Ocaml module provides the string interning backend necessary
for Symbols.v, but it cannot be used to bypass abstraction.

The new structure also makes it easy to switch out the implementation
of the Ident module in the future, or provide an alternative one, for
instance if we want to implement a version of Ident where of_string
and to_string can compute within Coq, sacrificing performance.