Reject casts to struct/union types (#68)
The ISO C99 standard allows cast only if the type name involved
is either a void type or a scalar type.  
For compatibility with GCC and Clang we used to support casting
a struct or union to exactly the same struct or union type.
That does not seem useful in practice and complicates conformance
testing.  This commit gets rid of this exception to the C99 rule.

Bug 23310