Bitfields transformation: record the fields after transformation
For each struct or union that contains bitfields,
we record the list of members after transformation,
where bit fields were replaced by carrier fields.

Right now we do not use this information but it will come handy
to fix a problem with struct initialization.

Also: clear the global hash tables on entry so that multiple runs
of the Bitfields transformation don't interfere with each other.
There probably was no interference before because identifiers are unique
enough, but this is fragile.