Record value of constant expression in C.Scase constructor
The Elab pass checks that the argument of 'case' is a compile-time constant
expression.  This commit records the value of this expression in the
C.Scase AST generated by Elab, so that it can be used for further
diagnostics, i.e. checking (in Elab) for duplicate cases.

Note that C2C ignores the recorded value and recomputes the value of
the expression using Ceval.integer_expr.  This is intentional:
Ceval.integer_expr is more trustworthy, as it is formally verified
against the CompCert C semantics.